renderers: make render_count calculation in SurfaceAnim functions unique and precise

+++ src/renderer/tr_animation_mds.c
@@ -1340,12 +1340,12 @@ void RB_SurfaceAnim(mdsSurface_t *surface)
 		{
 			lodScale = 0.35;
 		}
-		render_count = ROUND_INT(surface->numVerts * lodScale);
+		render_count = ROUND_INT((float) surface->numVerts * lodScale);
 
 	}
 	else
 	{
-		render_count = ROUND_INT(surface->numVerts * lodScale);
+		render_count = ROUND_INT((float) surface->numVerts * lodScale);
 		if (render_count < surface->minLod)
 		{
 			if (!(refent->reFlags & REFLAG_DEAD_LOD))
